+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make BareosDirPluginPrometheusExporter.py work with python3 [PR #1647]
 - Improve FreeBSD dependencies [PR #1670]
 - python-bareos: integrate usage of config files [PR #1678]
+- cmake: cleanup [PR #1661]
 
 ### Removed
 - plugins: remove old deprecated postgres plugin [PR #1606]
@@ -69,6 +70,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR #1655]: https://github.com/bareos/bareos/pull/1655
 [PR #1656]: https://github.com/bareos/bareos/pull/1656
 [PR #1659]: https://github.com/bareos/bareos/pull/1659
+[PR #1661]: https://github.com/bareos/bareos/pull/1661
 [PR #1665]: https://github.com/bareos/bareos/pull/1665
 [PR #1670]: https://github.com/bareos/bareos/pull/1670
 [PR #1671]: https://github.com/bareos/bareos/pull/1671